requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0

# Database
psycopg[binary]==3.2.3  # PostgreSQL driver (v3, better compatibility)

# ADB & Image Processing
pure-python-adb==0.3.0
Pillow==10.0.1
pytesseract==0.3.10

# HTTP Requests (for TrueMoney API)
requests==2.31.0

# Production Server
gunicorn==21.2.0
